deployment/requirements.txt
google-api-core==1.10.0
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.3.0
google-cloud==0.34.0
google-cloud-bigquery==1.11.2
google-cloud-core==0.29.1
google-cloud-datastore==1.7.3
google-cloud-storage==1.15.0
googleapis-common-protos==1.5.10
grpcio==1.20.1
github3.py==1.3.0